Rename RepositoryCard to Repository
