+++ apps/charterafrica/src/components/Repository/index.js
@@ -1,0 +1,3 @@
+import Repository from "./Repository";
+
+export default Repository;
