--- apps/charterafrica/src/components/RepositoryCard/index.js
@@ -1,3 +1,0 @@
-import RepositoryCard from "./RepositoryCard";
-
-export default RepositoryCard;
